change WeMath file & build_prompt (#755)
